1、jwt 引入redis 进行token管理，token续期 2、upms api

+++ ltts-common/ltts-common-jwt/src/main/java/com/lming/ltts/common/jwt/config/JwtProperties.java
@@ -43,7 +43,7 @@ public class JwtProperties {
     /**
      * 过期毫秒数
      */
-    private int expiresSecond = 1000 * 60 * 30;
+    private Long expiresSecond = 1000 * 60 * 30L;
     /**
      * 前端token前缀
      */
