+++ ltts-common/ltts-common-jwt/src/main/java/com/lming/ltts/common/jwt/util/JwtTokenUtil.java
@@ -76,7 +76,7 @@ public class JwtTokenUtil {
                     .setAudience(jwtProperties.getName())          // 代表这个JWT的接收对象；
                     .signWith(signatureAlgorithm, signingKey);
             //添加Token过期时间
-            int TTLMillis = jwtProperties.getExpiresSecond();
+            Long TTLMillis = jwtProperties.getExpiresSecond();
             if (TTLMillis >= 0) {
                 long expMillis = nowMillis + TTLMillis;
                 Date exp = new Date(expMillis);
